feat: 네비게이터 및 스크린 설정
